:truck: Move hero interface from feature component to app root
